requirements.txt
fastapi
uvicorn[standard]
pyyaml
httpx>=0.25.0
python-dotenv>=1.0.0
